refactor: Extract class tftp::base to new files tftpBase.*
u-t OK! 1834041 checks

Github issue #4
