+++ source/tftp_server.cpp
@@ -20,6 +20,7 @@
 #include <unistd.h>
 
 #include "tftp_server.h"
+#include "tftp_common.h"
 
 namespace tftp
 {
